Switch to RBR for Gerrit's MySQL binlogs
MySQL binlogs are filling up quick due to lots of delete statements during user
sync that don't delete any rows.  We could test to see if the rows exist first
but this would be a performance hit on a script that takes forever to run as it
is (plus the new sync script will work differently and not have this problem).

Switching to RBR which will only log row changes so will ignore these deletes.

Fixes bug #1002550

Change-Id: Ief2097de63bd0a5bde129b72cec87ff3745a6ea1