Make sure Gerrit sites are using the proper libraries.
The newer version of Bouncy Castle on Trusty requires an
additional package (libbbcpkix-java).  This package is
not available in ubuntu precise repository.

This change makes sure we are installing and using the
equivalent Bouncy Castle and MySQL libs from the ubuntu
package repo.

Notes:
Libs have been tested on review-dev.o.o (trusty) and
seem to work with both Gerrit 2.8.4 and 2.9.4 sites.

Change-Id: Ia5bfbc7b2f8c2148e7379d600dcc884dd35c624d