Fix serving of gerrit repos with '/p' in url
This change is similiar to I70d814a5eae
By default /p is served by gerrit itself so the normal proxying to
gerrit makes /p work. However for us we want /p to bypass gerrit
and hit the local git repo replica which requires running a git cgi
script so apache needs to be told that it is ok to do that.

It was giving the following error:
'fatal: unable to access 'https://review-dev.openstack.org/p/gtest-org/test2/':
The requested URL returned error: 403'

Change-Id: Iffdcefb2215a468fce3836d46377f4d1344e73b9