Run lib tidy after plugin install, before start
Ensure that we run the java lib tidy step after we have installed our
Gerrit plugins and before we start the Gerrit service. This ensures that
all gerrit init and plugin isntalls are done which may install libs to
review_site/lib that need to be cleaned up before we start gerrit. The
tidy step is what we use for this cleanup so it should happen in this
order.

Change-Id: I15b69e9da17640cb8b7ce31f59edf63395c6785d