Setup preconditions for acceptance tests.
We split the definition of the module under test from its preconditions.

As part of it, we have identified that we were installing the wrong pip
(using python-pip package). The python-pip package has a version that is
incompatible with puppet package provider. We changed it to install pip
using easy_install utility, which is compatible with puppet.

This behaviour was identified while testing puppet-zuul. To keep it
consistent across projects, we decided to also apply the same logic on
this module.

Change-Id: I0b9014ce6470d5c142ab6a3b281ea7401f88e738
Co-Authored-By: Bruno Tavares <btavare@thoughtworks.com>
Co-Authored-By: Danilo Ramalho <dramalho@thoughtworks.com>