Reduce time to reindex Gerrit data
Reindexing a Gerrit db can be quite slow.  It currently takes approximately
50 minutes to reindex review.o.o on an HP 'standard.4xlarge' VM (60 GB RAM
and 12 VCPUs) with the data stored on a local MySQL db.  Allowing Gerrit to
use more threads will make it run much faster.  For example, using 4 threads
will cut the time down to 15 mins.  This change sets the default number of
threads to half the number of system CPUs.

Change-Id: Ia35a6b4ab725b6c38d04c5de6975622f21cc11db