Allow setting the number of threads for indexing Gerrit changes
Indexing Gerrit changes can be a time consuming operation.  This
change allows us to set the number of threads to use for indexing
operation which will hopefully make the indexing faster.  Set the
default number of threads to the same as the number of threads
used for reindexing.  This option is available in Gerrit 2.8

referenced in index.threads section:
 https://review.openstack.org/Documentation/config-gerrit.html

Change-Id: I3aa1909d3edc8aa52d026b3f8e0dfafac4e14f6d