301 Redirect stripping out double leading slash
In the case of a double leading slash (as is seen in 2.11.x) due to
issue https://code.google.com/p/gerrit/issues/detail?id=3365 perform
a mod_rewrite that 301 redirects and strips the preceeding slash.

In this case the only reliable way to handle the slash strip is to
directly act upon %{THE_REQUEST} because mod_rewrite does not see the
leading slashes in the %{REQUEST_URI} itself and therefore cannot be
stripped with a simple "RewriteRule ^//(.*) ...".

The root of this issue is related to how gerrit decodes and mod_rewite
interacts with the decoded/encoded URLs while passing the referrer off
to the OpenId SSO provider. This is a hack-y work around until the
underlying issue can be resolved. This should have no impact on any
requests that do not have a double leading slash.

Change-Id: Ic2b685960c340cc458caac3f2ebb28407b271bae