Update Gerrit proxy configuration.
Replace the apache reverse proxy configs with the recommended
settings from the gerrit documentation[1].

Explicit use of ProxyPass <URL> ! is used to handle cases where
we do not want to proxy through to gerrit (such as for '/p')
where an offload is expected/desired

RewriteRule was causing the double slashes in Gerrit url (i.e.
https://review-dev.openstack.org//#/q/status:open). Removing it
fixes this issue.  More explanation can be found in gerrit
upstream bug[2].

[1] https://gerrit-review.googlesource.com/Documentation/
    config-reverseproxy.html#_apache_2_configuration
[2] https://code.google.com/p/gerrit/issues/detail?id=3365

Change-Id: Ied8ce7a0389253c10225a8bf3671f09a879bb061
Co-Authored-By: Khai Do <zaro0508@gmail.com>