Allow puppet to enabling or disabling Gerrit drafts feature
In Gerrit 2.11 disabling draft pushes in gerrit ACLS was not enought to
disable the drafts feature.  There is now an change.allowDrafts [1]
settting in gerrit.config to disable drafts which also needs to be
set. Gerrit enables drafts by default but we want it disabled
because we prefer using the 'work in progress' (WIP) workflow.

[1] https://review-dev.openstack.org/Documentation/config-gerrit.html#change

depends-on: I056720e79a44587bee60169f1c33403a63b78e7a
Change-Id: Iea970a85d433d8707102a3dcaf8cbae2d5662b6e