Change line lengths for commit message warnings
Set maximum length for subject line to 50, default is 65
Set maximum length for message lines to 72, default is 70.

Explicitly specify that it is only a warning if these lengths are
exceeded and not an error.

The values chosen are based on the OpenStack guidelines:
https://wiki.openstack.org/wiki/GitCommitMessages#Summary_of_Git_commit_message_structure

Configuration information for the Gerrit
commit-message-length-validator plugin:
https://gerrit.googlesource.com/plugins/commit-message-length-validator/+doc/master/src/main/resources/Documentation/config.md

Change-Id: If871518d8d667a1b98d88c7fd4edecd31dd0fc33