Fix gerrit plugin installation
The current test to only check for the existance of a plugin file
doesn't work on update scenarios.  The file already exists so it
won't run the install.  Instead we should download the file if it
doesn't already exist and then install after download.

This should install a plugins in 2 cases:
  1. install plugin if there isn't already one in place
  2. install an updated plugin if a new version of the plugin has
     been downloaded

Change-Id: Id27038a91c332650404058aa5d6c461f00c92592