Add configurations for gerrit its plugins
Add configurations section for its plugins.  Settings in
gerrit.configi and secure.config is to associate gerrit with an its
and settings in actions.config is to define the actions that the
plugin should take when a gerrit change gets updated. These
configurations are provided by the its-base plugin[1].  This change
will allow puppet to configure the its-storyboard plugin.

[1] https://gerrit.googlesource.com/plugins/its-base/+/stable-2.11/src/main/resources/Documentation/config-common.md
Change-Id: I9f47a2ed88ffbe827e494a478c0dc89a08bbe370