run git gc instead of repack for git repos
Run git gc instead of repack on all repos.  Testing[1] indicates that
gc-ing reduces repo size and helps maintain better performance for
apps that host the repos.

Change I4e171b17a7b4 replaced repack with gc for the git replication
and mirrors but not for the repos in the gerrit site.

[1] http://lists.openstack.org/pipermail/openstack-infra/2016-June/004403.html

Change-Id: I7a5364a07c3e07ffc9a053a73fab0c16861fc1c5