Refactor how puppet sets Gerrit's gitweb configs
Consolidate gitweb and cgit parameters into one gitweb parameter
containing a dictionary of gitweb settings.  This will make it
easier to add and update gitweb parameters as Gerrit's support
for gitweb[1] evolves.

[1] https://review.openstack.org/Documentation/config-gerrit.html#gitweb

TODO: keeping cgit, web_repo_url, and web_repo_url_encode
in init.pp to keep puppet happy (as a transition step).
Those params are not used and should be removed in a
follow up change.  Also needed to handle gitweb as both
boolean and hash types.  The boolean type should be remove
as part of the cleanup.

Change-Id: Ie470a38bc14a2b37a43f45be274ed193c1c00a1a