Make gerrit work on xenial
First, add a params class so that we can distinguish between java packages
on different Ubuntu releases.

Set sql-mode for mysql to remove NO_ZERO_DATE mode so gerrit can
initialize. If not removed, gerrit tries to create the
account_group_by_id_aud table with column added_on with a default value
of 0 and fails. This is only important for CI and downstream since
Infra's gerrit instance uses trove instead of a local database.

Depends-On: If25da754ec107b8dba2d9343f16b54b99cc88e01
Change-Id: Ifd2244ae9dd212b2475f9cd6adb994bc058a4769