Make beaker tests work on xenial
Remove all spec assertions for the existence of files. We can assume
that if the puppet finished successfully then the config files were also
laid down correctly. We remove these checks instead of updating the test
for /home/gerrit/id_rsa.pub which was moved to
/home/gerrit/.ssh/id_rsa.pub.

Remove all spec assertions for the existence of packages. Again, we can
assume that if puppet finished successfully then the package was found
and installed. We remove these checks instead of updating the test for
the JRE package to point to the new Xenial package.

Don't set java_home in the gerrit class since the module should be
setting the right default for us based on the release.

Make mysql config changes prompt a service restart (only in the
fixture). gerrit::mysql sets sql-mode to remove NO_ZERO_DATE, but the
mysql module doesn't restart the service by default when the config
changes and bindep preinstalls mysql so the service is already started
when my.cnf is set.

Update the beaker nodeset and spec helper to run on xenial and install
puppet from the Ubuntu repos instead of from puppetlabs since puppetlabs
doesn't support puppet 3 on xenial.

Change-Id: I0b6db86a76fc532443f80d6cbed2b1ef6ca9605e