Support setting accountPatchReviewDb.url in config
Add accountPatchReviewDb.url as a configurable option to the secure
config file (we put it in secure config as the url likely contains a
database password in it). This is a new database in Gerrit 2.13 which
will default to using an H2 db in review_site/db if a url is not
specified.

Note that this database must have a different name than the review
database (called 'reviewdb' by this module) and if they are the same
that database will be dropped and you will have a sad afternoon.

Change-Id: Ifc04395f076200a68e6398190a0712c80e5278c6