Fix beaker tests
On puppet 4, the tests were failing because the new ordering algorithm
caused the gerrit initializer exec resource to happen before the
database was set up. This patch adds an explicit ordering to the
gerrit::mysql class in the test fixtures to ensure the database is set
up first. This shouldn't be an issue in real life because we don't use
puppet to set up the database.

Also override the vhost_name and canonical server name parameters with
'localhost' since the $fqdn fact on test nodes can be wonky without some
extra setup.

Also add a conditional to the ssl-cert group exec so that it doesn't run
every time. Unfortunately this is not enough to make the tests
idempotent since the tidy resource behaves strangely when the directory
contains a symlink.

Change-Id: I56174d98496cd640b5b6490650a643c1aed740f7