Set java lib user ownership to gerrit2
The Gerrit init process likes to rename these libs and it cannot do so
when they are owned by root. Update the process to download and install
the libs out of band of the init process to run as gerrit2 so that these
libs are owned by the gerrit2 user.

Change-Id: I1b619a64c5e324fa8d0c24ef02ace9561c382475