Revert "Create index on fresh install"
This reverts commit 5c44d09dc717068f7ea5ee0ff0d2319863991817.

This change forced us to do an offline reindex whenever the war was
updated. Unfortunately, that means even on minor upgrades we do a long
reindex, which times out in puppet, then puppet fails as does an
upgrade.

You can address this by setting offline_reindex = true when doing a new
install, then flip that to false (the default) once the new install is
bootstrapped.

Change-Id: I1f730a1cd2bcf96e6b94f2d0a82078f46d49e0ce